+++ src/VisualStudio/IntegrationTest/IntegrationTests/AbstractEditorTest.cs
@@ -9,7 +9,6 @@ using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Microsoft.VisualStudio.IntegrationTest.Utilities;
 using Microsoft.VisualStudio.IntegrationTest.Utilities.Common;
-using Microsoft.VisualStudio.IntegrationTest.Utilities.InProcess;
 using Microsoft.VisualStudio.IntegrationTest.Utilities.Input;
 using Microsoft.VisualStudio.IntegrationTest.Utilities.OutOfProcess;
 using Roslyn.Test.Utilities;
@@ -138,15 +137,6 @@ namespace Roslyn.VisualStudio.IntegrationTests
         protected void SendKeys(params object[] keys)
             => Editor.SendKeys(keys);
 
-        protected KeyPress KeyPress(VirtualKey virtualKey, ShiftState shiftState)
-            => new KeyPress(virtualKey, shiftState);
-
-        protected KeyPress Ctrl(VirtualKey virtualKey)
-            => new KeyPress(virtualKey, ShiftState.Ctrl);
-
-        protected KeyPress Shift(VirtualKey virtualKey)
-            => new KeyPress(virtualKey, ShiftState.Shift);
-
         protected void DisableSuggestionMode()
             => VisualStudioWorkspaceOutOfProc.SetUseSuggestionMode(false);
 
@@ -174,9 +164,6 @@ namespace Roslyn.VisualStudio.IntegrationTests
             Editor.WaitForLightBulbSession();
             WaitForAsyncOperations(FeatureAttribute.LightBulb);
         }
-
-        protected void ExecuteCommand(string commandName, string argument = "")
-            => VisualStudio.Instance.ExecuteCommand(commandName, argument);
 
         private void VerifyCurrentLineTextAndAssertCaretPosition(string expectedText, bool trimWhitespace)
         {
