+++ src/VisualStudio/IntegrationTest/IntegrationTests/AbstractIntegrationTest.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Threading;
 using Microsoft.VisualStudio.IntegrationTest.Utilities;
+using Microsoft.VisualStudio.IntegrationTest.Utilities.Input;
 
 namespace Roslyn.VisualStudio.IntegrationTests
 {
@@ -24,5 +25,20 @@ namespace Roslyn.VisualStudio.IntegrationTests
             var timeout = TimeSpan.FromMilliseconds(seconds * 1000);
             Thread.Sleep(timeout);
         }
+
+        protected KeyPress KeyPress(VirtualKey virtualKey, ShiftState shiftState)
+            => new KeyPress(virtualKey, shiftState);
+
+        protected KeyPress Ctrl(VirtualKey virtualKey)
+            => new KeyPress(virtualKey, ShiftState.Ctrl);
+
+        protected KeyPress Shift(VirtualKey virtualKey)
+            => new KeyPress(virtualKey, ShiftState.Shift);
+
+        protected KeyPress Alt(VirtualKey virtualKey)
+            => new KeyPress(virtualKey, ShiftState.Alt);
+
+        protected void ExecuteCommand(string commandName, string argument = "")
+            => VisualStudio.Instance.ExecuteCommand(commandName, argument);
     }
 }
